Fixing issue with loading plugin
The plugin would load and show as loaded but an error while loading the version file causes the plugin to silently crash and not trigger portal events. The issue was discovered when updating the server from 1.7.10 to 1.8, but is still present in 1.12.2.
This commit uses an alternative method that has worked fine in production for years.